Tracking many datapoints with approx averages
This diff introduces a new class that will maintain a circular array of information about the interrupts that are fired and when.  It then supports a generic mechanism to get the number of interrupts that have occurred for arbitary (but bounded by unsigned long in -- about 1 hour) time periods.

This is a first pass and still needs debugging and testing.