Working average counter
This is a complete implementation of an "duration adaptive average calculator".   It is a container class where you add items to it along with the time that the items were seen.  Then you can ask for an average value from a given point in time and for a specific duration.  If we have sufficient datapoints to cover the duration (as defined by the oldest datapoint and the duration of time) then we compute the average.

The data structure is maintained as a circular array of fixed size.